Port to 1.18.2 and fixed #1

+++ src/main/java/com/black_dog20/torchaction/common/events/EventHandler.java
@@ -11,6 +11,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraftforge.event.entity.player.EntityItemPickupEvent;
+import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.items.CapabilityItemHandler;
@@ -34,6 +35,15 @@ public class EventHandler {
             if (!TorchHolderProperties.getAutoPickupMode(torchHolder))
                 return;
 
+            boolean hasRoom = torchHolder.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
+                    .filter(TorchItemHandler.class::isInstance)
+                    .map(TorchItemHandler.class::cast)
+                    .map(TorchItemHandler::hasRoom)
+                    .orElse(false);
+
+            if (!hasRoom)
+                return;
+
             torchHolder.getCapability(CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)
                     .filter(TorchItemHandler.class::isInstance)
                     .map(TorchItemHandler.class::cast)
@@ -49,9 +59,15 @@ public class EventHandler {
             }
             if (remainingStackSimulated.get().isEmpty()) {
                 itemEntity.setItem(ItemStack.EMPTY);
-                event.setCanceled(true);
+                event.setResult(Event.Result.ALLOW);
             } else {
-                itemEntity.setItem(remainingStackSimulated.get());
+                if (player.getInventory().add(remainingStackSimulated.get())) {
+                    itemEntity.setItem(ItemStack.EMPTY);
+                    event.setResult(Event.Result.ALLOW);
+                } else {
+                    itemEntity.setItem(remainingStackSimulated.get());
+                    event.setCanceled(true);
+                }
             }
         }
